+++ src/XUnity.AutoTranslator.Plugin.Core/UI/ToggleViewModel.cs
@@ -8,12 +8,13 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
       private GUIContent _enabled;
       private GUIContent _disabled;
 
-      public ToggleViewModel( string text, string enabledTooltip, string disabledTooltip, Action onToggled, Func<bool> isToggled )
+      public ToggleViewModel( string text, string enabledTooltip, string disabledTooltip, Action onToggled, Func<bool> isToggled, bool enabled = true )
       {
          _enabled = new GUIContent( text, enabledTooltip );
          _disabled = new GUIContent( text, disabledTooltip );
          OnToggled = onToggled;
          IsToggled = isToggled;
+         Enabled = enabled;
       }
 
       public GUIContent Text
@@ -27,6 +28,8 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
             return _disabled;
          }
       }
+
+      public bool Enabled { get; set; }
 
       public Action OnToggled { get; set; }